Expanded translation of Decl, Type and Stmt classes to Python.
This is still very much a work in progress, and the Cython code is quite messy.
I have updated the 'test/test_parser.py' test suite to parse and extract
information about a function declaration/definition.

At this stage we can get the function name, parameters and result type. Next
up is the "body", which means translating a load of Stmt subclasses.